+++ src/flask/typing.py
@@ -2,8 +2,8 @@ import typing as t
 
 
 if t.TYPE_CHECKING:
+    from _typeshed.wsgi import WSGIApplication  # noqa: F401
     from werkzeug.datastructures import Headers  # noqa: F401
-    from wsgiref.types import WSGIApplication  # noqa: F401
     from .wrappers import Response  # noqa: F401
 
 # The possible types that are directly convertible or are a Response object.